refactor : Refactor codes based on Jihoon's review (4)
