+++ src/main/java/lotto/lottogenerator/LottoNumGeneratorByAuto.java
@@ -3,15 +3,11 @@ package lotto.lottogenerator;
 import java.util.Collections;
 import java.util.List;
 
-public class LottoNumGeneratorByAuto implements LottoNumGeneratorStrategy {
+public class LottoNumGeneratorByAuto {
     private static final int LOTTO_SIZE = 6;
 
-    public LottoNumGeneratorByAuto() {
-    }
-
-    @Override
-    public List<Integer> generate() {
-        List<Integer> lottoNumbers = LottoNumberPool.makePool();
+    public static List<Integer> generate() {
+        List<Integer> lottoNumbers = LottoNumMaker.makePool();
         Collections.shuffle(lottoNumbers);
         lottoNumbers = lottoNumbers.subList(0, LOTTO_SIZE);
         Collections.sort(lottoNumbers);
