--- src/main/java/lotto/lottogenerator/LottoNumGeneratorStrategy.java
@@ -1,7 +1,0 @@
-package lotto.lottogenerator;
-
-import java.util.List;
-
-public interface LottoNumGeneratorStrategy {
-    List<Integer> generate();
-}
