+++ src/main/java/lotto/lottogenerator/LottoNumMaker.java
@@ -1,0 +1,19 @@
+package lotto.lottogenerator;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class LottoNumMaker {
+    private static final int LOTTO_NUMBER_MIN = 1;
+    private static final int LOTTO_NUMBER_MAX = 45;
+
+    public static List<Integer> makePool() {
+        List<Integer> pool = new ArrayList<>();
+
+        for (int i = LOTTO_NUMBER_MIN; i <= LOTTO_NUMBER_MAX; i++) {
+            pool.add(i);
+        }
+
+        return pool;
+    }
+}
