+++ src/test/java/lotto/model/LottoNumbersTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import lotto.model.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
