+++ src/test/java/lotto/model/MoneyTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import lotto.model.Money;
 import org.junit.jupiter.api.DisplayName;
